google-adk/weather_agent/requirements.txt
litellm
google-adk
dotenv
